Finish extension ports
