+++ PrincessPaperplane/.local.env
@@ -2,18 +2,9 @@
 
 # Discord Server IDs
 DISCORD.SERVER.LIVE=
-DISCORD.SERVER.TEST=
 
 # Discord Role Channel IDs
 DISCORD.CHANNEL.ROLE.LIVE=
-DISCORD.CHANNEL.ROLE.TEST=
 
 # Discord Level Channel IDs
 DISCORD.CHANNEL.LEVEL.LIVE=
-DISCORD.CHANNEL.LEVEL.TEST=
-
-# MODULE SPECIFIC SECRETS
-# Twitter API Secrets
-TWITTER.BEARER.ART=
-TWITTER.CHANNEL.CONFIRM=
-TWITTER.TIMER=
