+++ PrincessPaperplane/extension/twitter/configuration/.default
@@ -1,0 +1,3 @@
+TWITTER.BEARER.ART=
+TWITTER.CHANNEL.CONFIRM=
+TWITTER.TIMER=
