+++ PrincessPaperplane/resources/database-dumps/princesspaperplane_quotes.sql
