+++ PrincessPaperplane/utility/log.py
@@ -1,5 +1,4 @@
 import logging
-from typing import Any
 
 
 def setup():
